Revert "Move installBuiltInExtensions out of the ExtensionList Class (#1827)"
This reverts commit fc47b25afd94a9bb27c0ec6a227bece749aedcea.
